requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.8.5
aiosignal==1.3.2
altgraph==0.17.4
async-timeout==4.0.3
asyncio==3.4.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.3.0
cryptography==41.0.3
cssselect2==0.8.0
cycler==0.12.1
dnspython==2.4.2
et_xmlfile==2.0.0
fonttools==4.58.0
frozenlist==1.6.0
future==1.0.0
idna==3.10
importlib_metadata==8.7.0
ipaddress==1.0.23
kiwisolver==1.4.7
matplotlib==3.6.0
multidict==6.4.3
nest-asyncio==1.6.0
netifaces==0.11.0
# Editable install with no version control (nettools==1.0.0)
-e c:\users\administrator\desktop\nettools
networkx==3.2.1
numpy==1.23.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
paramiko==3.5.1
pdfkit==1.0.0
pefile==2023.2.7
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydyf==0.11.0
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.3
pyphen==0.17.2
pyping==0.0.6
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-nmap==0.7.1
python-whois==0.8.0
pytz==2025.2
pywin32==305
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==6.1.0
reportlab==4.4.0
requests==2.31.0
scapy==2.5.0
scipy==1.13.1
six==1.17.0
soupsieve==2.7
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.66.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
weasyprint==65.1
webencodings==0.5.1
winshell==0.6
wxPython==4.2.1
XlsxWriter==3.2.3
yarl==1.20.0
zipp==3.21.0
zopfli==0.2.3.post1
